requirements.txt
mysql-connector-python-rf
mysql-connector-python~=8.1.0
django~=4.2.6
bs4~=0.0.1
beautifulsoup4~=4.12.2